creating and deleting service files now works aswell

+++ prefs.js
@@ -1,11 +1,13 @@
 'use strict';
 
-const { Adw, Gio, Gtk } = imports.gi;
+const { Adw, Gio, Gtk, GLib } = imports.gi;
 
+const Bytes  = imports.byteArray
 const ExtensionUtils = imports.misc.extensionUtils;
 const Me = ExtensionUtils.getCurrentExtension();
 
-const Utils             = Me.imports.utils
+const Utils             = Me.imports.utils;
+const Service           = Me.imports.service;
 
 function init() {
 }
@@ -91,7 +93,7 @@ function fillPreferencesWindow(window) {
 function fillEditTunnelPage(window,tunnel=null){
     let edit=true;
     if (tunnel == null) {
-        tunnel = new Utils.Tunnel(defaultHost,[],[],defaultOptions)
+        tunnel = new Utils.Tunnel(defaultHost,[],[],defaultOptions,getCurrentUser())
         edit=false;
     }
     const editTunnelPage = new Adw.PreferencesPage();
@@ -113,6 +115,11 @@ function fillEditTunnelPage(window,tunnel=null){
     const editTunnelGroup = new Adw.PreferencesGroup();
     editTunnelGroup.set_title("Tunnel Properties")
     editTunnelPage.add(editTunnelGroup);
+
+    const tunnelUser = new Adw.EntryRow();
+    tunnelUser.set_title("Local User to execute Service");
+    tunnelUser.set_text(tunnel.user);
+    editTunnelGroup.add(tunnelUser);
 
     const tunnelHost = new Adw.EntryRow();
     tunnelHost.set_title("SSH Host");
@@ -140,13 +147,16 @@ function fillEditTunnelPage(window,tunnel=null){
     }
     confirmButton.connect('clicked', () => {
         tunnel.hostname=tunnelHost.get_text();
+        tunnel.user=tunnelUser.get_text();
         tunnel.forwards=getArrayFromRows(forwardsRows);
         tunnel.reverses=getArrayFromRows(reversesRows);
         tunnel.options=getArrayFromRows(optionsRows);
 
-        saveTunnel(tunnel);
-        refreshSettingsPage(window);
-        window.close_subpage();
+        if (createServiceFiles(tunnel)){
+            saveTunnel(tunnel);
+            refreshSettingsPage(window);
+            window.close_subpage();
+        }
     });
 
     if (edit){
@@ -157,9 +167,11 @@ function fillEditTunnelPage(window,tunnel=null){
         deleteTunnelButton.set_icon_name("");
         deleteTunnelButton.set_label("Delete Tunnel");
         deleteTunnelButton.connect('clicked', () => {
-            deleteTunnel(tunnel);
-            refreshSettingsPage(window);
-            window.close_subpage();
+            if (deleteServiceFiles(tunnel)) {
+                deleteTunnel(tunnel);
+                refreshSettingsPage(window);
+                window.close_subpage();
+            }
         });
         deleteTunnelGroup.add(deleteTunnelButton);
     }
@@ -243,6 +255,122 @@ function deleteTunnel(tunnel){
     tunnels.splice(id, 1); // 2nd parameter means remove one item only
     let str = stringifyTunnels(tunnels);
     settings.set_strv("tunnels",str);
+}
+
+function createServiceFiles(tunnel){
+    let files = [];
+    let serviceFile = getServiceFile(tunnel);
+    if (serviceFile != null){
+        files.push(serviceFile);
+    }
+    let tunnelFile = getTunnelFile(tunnel);
+    if (tunnelFile != null){
+        files.push(tunnelFile);
+    }
+    return moveFiles(files);
+}
+
+function getServiceFile(tunnel){
+    const serviceFile = Gio.File.new_for_path('/etc/systemd/system/sshtunnel.' + tunnel.user + '@.service');
+    let replacements = {"%USER%":tunnel.user};
+    let str = Service.getServiceFile().replace(/%\w+%/g, function(all) {
+        return replacements[all] || "";
+    });
+
+    if (serviceFile.query_exists(null)){
+        if (checkFileContent(serviceFile.get_path(),str)) {
+            return null;
+        }
+    }
+    const [tempFile,] = Gio.File.new_tmp(null);
+    const res = tempFile.replace_contents(str, null, false, Gio.FileCreateFlags.REPLACE_DESTINATION, null);
+    return [tempFile.get_path(),serviceFile.get_path()];
+}
+
+function getTunnelFile(tunnel){
+    const tunnelFile = Gio.File.new_for_path('/etc/default/sshtunnel@' + tunnel.id);
+    let forwards = getStringFromArray(tunnel.forwards,"-L");
+    let reverses = getStringFromArray(tunnel.reverses,"-R");
+    let options = getStringFromArray(tunnel.options,"-o");
+    let replacements = {"%HOSTNAME%":tunnel.hostname,"%FORWARDS%":forwards,"%REVERSES%":reverses,"%OPTIONS%":options};
+    let str = Service.getTunnelFile().replace(/%\w+%/g, function(all) {
+        return replacements[all] || "";
+    });
+
+    if (tunnelFile.query_exists(null)){
+        if (checkFileContent(tunnelFile.get_path(),str)) {
+            return null;
+        }
+    }
+    const [tempFile,] = Gio.File.new_tmp(null);
+    const res = tempFile.replace_contents(str, null, false, Gio.FileCreateFlags.REPLACE_DESTINATION, null);
+    return [tempFile.get_path(),tunnelFile.get_path()];
+}
+
+function deleteServiceFiles(tunnel){
+    const settings = ExtensionUtils.getSettings(
+        'org.gnome.shell.extensions.sshtunnel');
+    let tunnels = parseTunnels(settings.get_strv("tunnels"));
+    let tunnelsSameUser = tunnels.filter(obj => obj.user == tunnel.user);
+    let files = ['/etc/default/sshtunnel@' + tunnel.id];
+    if (tunnelsSameUser.length = 1){
+        files.push('/etc/systemd/system/sshtunnel.' + tunnel.user + '@.service');
+    }
+    return deleteFiles(files);
+}
+
+function getStringFromArray(arr,delimiter){
+    let str = "";
+    arr.forEach(obj => {
+        str +=delimiter + " " + obj + " ";
+    });
+    return str;
+}
+
+function checkFileContent(path,text){
+    const file = Gio.File.new_for_path(path);
+    if (file.query_exists(null)){
+        const [, contents, etag] = file.load_contents(null);
+        const decoder = new TextDecoder('utf-8');
+        const contentsString = decoder.decode(contents);
+        if (text==contentsString){
+            return true;
+        }
+    }
+    return false;
+}
+
+function moveFiles(files){
+    if (files.length <= 0) {
+        return true;
+    }
+    let cmd = ""
+    files.forEach(obj => {
+        cmd += "mv " + obj[0] + " " + obj[1] + "; ";
+    })
+    cmd += "systemctl daemon-reload;"
+    const res = GLib.spawn_command_line_sync(`pkexec sh -c "${cmd}"`);
+    return res[3]==0;
+}
+
+function deleteFiles(files){
+    if (files.length <= 0) {
+        return true;
+    }
+    let cmd = "";
+    files.forEach(obj => {
+        cmd += "rm " + obj + "; ";
+    })
+    cmd += "systemctl daemon-reload;"
+    print(cmd);
+    let [ok, out, err, exit] = GLib.spawn_command_line_sync(`pkexec sh -c "${cmd}"`);
+    return exit == 0;
+}
+
+function getCurrentUser(){
+    let cmd = "whoami";
+    let [ok, out, err, exit] = GLib.spawn_command_line_sync(`${cmd}`);
+    return Bytes.toString(out).slice(0,-1);
 }
 
 function parseTunnels(str){
